etl: Move cache configuration to shared module
It will soon be used by an additional ETL process.
